# sources/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация по API",
        "description": "API документация",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/account/set-code": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Вход для кандидата.",
                "description": "Возвращение JSON объекта",
                "operationId": "65967bfcba29c1cec23328a2062d0aab",
                "requestBody": {
                    "description": "Данные для отправки кода",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "+7 (999) 999-99-99"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "УСПЕХ!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "phone": {
                                            "type": "string",
                                            "example": "+7 (999) 999-99-99"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Пользователь не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/accounts/store": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Создание аккаунта",
                "description": "Возвращение JSON объекта",
                "operationId": "05457050a2b2cfca5b836182193618b5",
                "requestBody": {
                    "description": "Данные для создания аккаунта",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "e8ff11fa-822b-11f0-8411-10f60a82b815"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Шихавцов"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Андрей"
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "example": "Александрович"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mail@mail.ru"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "УСПЕХ!"
                    }
                }
            }
        },
        "/api/v1/accounts/update": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Обновление аккаунта",
                "description": "Возвращение JSON объекта",
                "operationId": "4fc33b8eeba4fb7858058c814ac625d7",
                "requestBody": {
                    "description": "Данные для обновления аккаунта",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "key": {
                                        "type": "string",
                                        "example": "e8ff11fa-822b-11f0-8411-10f60a82b815"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Шихавцов"
                                    },
                                    "first_name": {
                                        "type": "string",
                                        "example": "Андрей"
                                    },
                                    "middle_name": {
                                        "type": "string",
                                        "example": "Александрович"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "mail@mail.ru"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "УСПЕХ!"
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Пользователь не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}